requirements.txt
pyrogram
python-dotenv
tqdm
flask
tgcrypto
